More /automatic updates.
Add a canvas cache to avoid re-rendering oscillators and gliders all the
time.  I'm not actually convinced this is a net performance win, but
it's neat.

Lock to 60fps.  This helps confuse me less when switching from a 120Hz
screen to a 60Hz one.

Add more information to the debugging overlay.